Added a new requests object.  This object keys on the requested URL. The values corresponding to the requested URL is an object containing the original options and the callback corresponding to that URL.
This change addresses a concurrency issue found in this library.  In the
previous version, calling jqoteload a second time before a previous
jqoteload had completed would cause the first success to operate with
the options and callback of the second request (which may or may not
have been the corresponding request) and the second success callback to
error.  This new version addresses this case.